move auto rewarder to base chain
