+++ rl_chain/__init__.py
@@ -1,10 +1,8 @@
 from .pick_best_chain import (
     PickBest,
-    PickBestAutoSelectionScorer,
 )
 from .slates_chain import (
     SlatesPersonalizerChain,
-    SlatesAutoSelectionScorer,
     SlatesRandomPolicy,
     SlatesFirstChoicePolicy,
 )
@@ -13,6 +11,7 @@ from .rl_chain_base import (
     BasedOn,
     ToSelectFrom,
     SelectionScorer,
+    AutoSelectionScorer,
     Embedder,
     Policy,
     VwPolicy,
