+++ rl_chain/pick_best_chain.py
@@ -2,12 +2,6 @@ from __future__ import annotations
 
 from . import rl_chain_base as base
 
-from langchain.prompts import (
-    ChatPromptTemplate,
-    SystemMessagePromptTemplate,
-    HumanMessagePromptTemplate,
-)
-
 from langchain.prompts.prompt import PromptTemplate
 
 from langchain.callbacks.manager import CallbackManagerForChainRun
@@ -81,45 +75,6 @@ class PickBestFeatureEmbedder(base.Embedder):
             example_string += "\n"
         # Strip the last newline
         return example_string[:-1]
-
-
-class PickBestAutoSelectionScorer(base.SelectionScorer):
-    llm_chain: LLMChain
-    prompt: PromptTemplate
-    default_system_prompt = SystemMessagePromptTemplate.from_template(
-        "PLEASE RESPOND ONLY WITH A SIGNLE FLOAT AND NO OTHER TEXT EXPLANATION\n You are a strict judge that is called on to rank a response based on given criteria.\
-                You must respond with your ranking by providing a single float within the range [-1, 1], -1 being very bad response and 1 being very good response."
-    )
-
-    def __init__(self, llm, prompt=None):
-        if prompt:
-            self.prompt = prompt
-        else:
-            human_template = 'Given this based_on "{best_pick_context}" as the most important attribute, rank how good or bad this text selection is: "{best_pick}".'
-            human_message_prompt = HumanMessagePromptTemplate.from_template(
-                human_template
-            )
-
-            chat_prompt = ChatPromptTemplate.from_messages(
-                [
-                    PickBestAutoSelectionScorer.default_system_prompt,
-                    human_message_prompt,
-                ]
-            )
-            self.prompt = chat_prompt
-
-        self.llm_chain = LLMChain(llm=llm, prompt=self.prompt)
-
-    def score_response(self, inputs: Dict[str, Any], llm_response: str) -> float:
-        ranking = self.llm_chain.predict(llm_response=llm_response, **inputs)
-        ranking = ranking.strip()
-        try:
-            resp = float(ranking)
-            return resp
-        except Exception as e:
-            raise RuntimeError(
-                f"The llm did not manage to rank the response as expected, there is always the option to try again or tweak the reward prompt. Error: {e}"
-            )
 
 
 class PickBest(base.RLChain):
@@ -300,7 +255,7 @@ class PickBest(base.RLChain):
 
     @property
     def _chain_type(self) -> str:
-        return "llm_rl_chain_pick_best_chain"
+        return "rl_chain_pick_best"
 
     @classmethod
     def from_chain(cls, llm_chain: Chain, prompt: PromptTemplate, **kwargs: Any):
